Remove empty arguments on Windows
Empty strings in the arguments list cause openscad
to fail on Windows (I think this is due to the way popen
converts from list to string).

This workaround gets rid of empty arguments.
It appears to fix the problem on Windows.
I don't think it breaks on other platforms as I've put it in
 an if statement